+++ services/director-v2/tests/unit/test_utils_substitutions_sessions.py
@@ -6,9 +6,7 @@
 
 
 import asyncio
-import inspect
 import json
-from typing import Any, Callable, NamedTuple, TypeAlias
 
 import pytest
 from faker import Faker
@@ -24,9 +22,12 @@ from simcore_service_director_v2.modules.dynamic_sidecar.docker_compose_specs_su
     substitute_session_environments,
     substitute_vendor_environments,
 )
-
-ContextDict: TypeAlias = dict[str, Any]
-ContextGetter: TypeAlias = Callable[[ContextDict], Any]
+from simcore_service_director_v2.utils.substitutions_sessions import (
+    ContextDict,
+    factory_context_getter,
+    factory_handler,
+    resolve_session_environs,
+)
 
 
 @pytest.fixture
@@ -43,49 +44,10 @@ def session_context(faker: Faker) -> ContextDict:
     )
 
 
-async def test_it(faker: Faker, session_context: ContextDict):
+async def test_resolve_session_environs(faker: Faker, session_context: ContextDict):
     assert substitute_session_environments
     assert substitute_vendor_environments
     assert substitute_request_environments
-
-    # Registered
-    class CaptureError(ValueError):
-        ...
-
-    def factory_context_getter(parameter_name: str) -> ContextGetter:
-        """Factory that creates a function that gets a context as argument and gets a named parameter
-
-        i.e. create_context_getter("foo")(context) == context["foo"]
-        """
-
-        def _get_or_raise(context: ContextDict) -> Any:
-            try:
-                return context[parameter_name]
-            except KeyError as err:
-                raise CaptureError(
-                    "Parameter {keyname} missing from substitution context"
-                ) from err
-
-        # For context["foo"] -> return operator.methodcaller("__getitem__", keyname)
-        # For context.foo -> return operator.attrgetter("project_id")
-        return _get_or_raise
-
-    class RequestTuple(NamedTuple):
-        handler: Callable
-        kwargs: dict[str, Any]
-
-    def factory_handler(coro: Callable) -> Callable[[ContextDict], RequestTuple]:
-        assert inspect.iscoroutinefunction(coro)  # nosec
-
-        def _create(context: ContextDict):
-            # NOTE: we could delay this as well ...
-            kwargs_from_context = {
-                param.name: factory_context_getter(param.name)(context)
-                for param in inspect.signature(request_user_role).parameters.values()
-            }
-            return RequestTuple(handler=coro, kwargs=kwargs_from_context)
-
-        return _create
 
     # ----------------
 
@@ -94,6 +56,8 @@ async def test_it(faker: Faker, session_context: ContextDict):
         "OSPARC_ENVIRONMENT_STUDY_UUID": factory_context_getter("project_id"),
         "OSPARC_ENVIRONMENT_NODE_UUID": factory_context_getter("node_id"),
     }
+
+    # Here some callbacks
 
     async def request_user_email(app: FastAPI, user_id: UserID) -> SubstitutionValue:
         print(app, user_id)
@@ -116,27 +80,7 @@ async def test_it(faker: Faker, session_context: ContextDict):
     # TODO: test validation errors handling
     # TODO: test timeout error handling
 
-    # prepares environs from context:
-    pre_environs: dict[str, SubstitutionValue | RequestTuple] = {
-        key: fun(session_context) for key, fun in oenvs_table.items()
-    }
-
-    # execute
-    environs: dict[str, SubstitutionValue] = {}
-
-    coros = {}
-    for key, value in pre_environs.items():
-        if isinstance(value, RequestTuple):
-            handler, kwargs = value
-            coro = handler(**kwargs)
-            # wraps to control timeout
-            coros[key] = asyncio.wait_for(coro, timeout=3)
-        else:
-            environs[key] = value
-
-    values = await asyncio.gather(*coros.values())
-    for key, value in zip(coros.keys(), values):
-        environs[key] = value
+    environs = await resolve_session_environs(oenvs_table, session_context)
 
     assert set(environs.keys()) == set(oenvs_table.keys())
